requirements.txt
# Paper A: Ghost Skill & Variance Collapse in PM10 Forecasting
# Dependencies for data processing, modeling, evaluation, and plotting

# Core scientific stack
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0

# Machine learning
lightgbm>=4.0.0
scikit-learn>=1.2.0

# Deep learning (LSTM)
tensorflow>=2.12.0

# Configuration & utilities
pyyaml>=6.0
joblib>=1.3.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
